requirements.txt
pytest==8.3.5
selene==2.0.0rc9
allure-pytest==2.14.2
selenium==4.31.0
requests==2.32.3
jsonschema